+++ packages/basic/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @grid-table/basic
 
+## 0.6.20
+
+### Patch Changes
+
+- @grid-table/core@0.6.20
+
 ## 0.6.19
 
 ### Patch Changes