+++ packages/example/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @grid-table/example
 
+## 0.0.33
+
+### Patch Changes
+
+- Updated dependencies
+  - @grid-table/view@0.6.20
+  - @grid-table/excel@0.5.23
+  - @grid-table/pivot@0.5.26
+  - @grid-table/core@0.6.20
+  - @grid-table/basic@0.6.20
+
 ## 0.0.32
 
 ### Patch Changes