+++ packages/excel/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @grid-table/excel
 
+## 0.5.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @grid-table/view@0.6.20
+  - @grid-table/core@0.6.20
+  - @grid-table/basic@0.6.20
+
 ## 0.5.22
 
 ### Patch Changes